+++ src/index.ts
@@ -17,3 +17,4 @@ export { TextInput } from './components/reusable/textInput';
 export { TextArea } from './components/reusable/textArea';
 export { ToggleButton } from './components/reusable/toggleButton';
 export { TimePicker } from './components/reusable/timepicker';
+export { Dropdown, DropdownOption } from './components/reusable/dropdown';